+++ include/shared/State.h
@@ -32,6 +32,7 @@ extern std::atomic<int> g_displayChangeGen;
 extern std::atomic<int> g_hudDecimalPlaces;
 extern std::atomic<bool> g_fortniteFocusedCache;
 extern std::atomic<bool> g_atomicShieldEnabled;
+extern std::atomic<bool> g_directHardwareModeEnabled;
 extern std::atomic<ULONGLONG> g_lastValidMatchTime;
 extern std::atomic<int> g_lockCount;
 extern std::atomic<bool> g_blockInputActive;